requirements.txt
dj-database-url==0.5.0
Django==2.2
gunicorn==19.9.0
psycopg2-binary==2.7.7
whitenoise==4.1.2
django-cors-headers==2.5.0
django-webpack-loader==0.6.0
python-decouple==3.1
djangorestframework==3.9.1
markdown==3.0.1
django-filter==2.1.0
Pillow==5.4.1
django-imagekit==4.0.2
xhtml2pdf==0.2.3
django-parler==1.9.2
django-parler-rest
